+++ src/Stormancer.Plugins/Limits/Stormancer.Server.Plugins.Limits/LimitsConfigurationBuilder.cs
@@ -1,0 +1,18 @@
+﻿namespace Stormancer.Server.Plugins.Limits
+{
+    /// <summary>
+    /// Limits configuration builder.
+    /// </summary>
+    public class LimitsConfigurationBuilder
+    {
+        /// <summary>
+        /// Gets the configuration.
+        /// </summary>
+        public LimitsConfiguration Config { get;  }
+
+        internal LimitsConfigurationBuilder(LimitsConfiguration config)
+        {
+            this.Config = config;
+        }
+    }
+}