+++ src/Stormancer.Plugins/GameSession/Stormancer.Server.Plugins.GameSession/Changelog.rst
@@ -9,10 +9,17 @@ Please use only reStructuredText in this file, no Markdown!
 
 This project adheres to semantic versioning.
 
-4.0.0.1
-----------
+4.0.1.1
+-------
 Changed
 *******
+- Declare GameSessionService at scene level.
+
+4.0.1
+-----
+Changed
+*******
+- Changed Behavior: do not disconnect users from the game session to allow them to restart a game.
 - Update to .NET5.0
 
 3.3.2.2