+++ src/Stormancer.Plugins/Party/Stormancer.Server.Plugins.Party/Changelog.rst
@@ -9,6 +9,12 @@ Please use only reStructuredText in this file, no Markdown!
 
 This project adheres to semantic versioning.
 
+4.0.0.2
+-------
+Changed
+*******
+- Update GameFinder dependency.
+
 4.0.0.1
 ----------
 Changed